requirements.txt
# Core RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.20
qdrant-client>=1.8.0
sentence-transformers>=2.5.0

# Document processing
pypdf>=4.0.0
pymupdf>=1.23.0
python-docx>=1.1.0
markdown>=3.5.0
unstructured>=0.18.0  # Updated for Python 3.12 support

# LLM integration
openai>=1.12.0

# Utilities
numpy>=1.26.0
pyyaml>=6.0.0
tqdm>=4.66.0

